# docs/storage-api.openapi.yaml
openapi: 3.1.0
info:
  title: Deniz Cloud Storage API
  version: 1.0.0
  description: |
    Cloud storage API for Deniz Cloud self-hosted home server.
    Provides file management (upload, download, rename, move, delete),
    folder CRUD, and resumable uploads via the TUS protocol.

    **Base URL:** `https://storage.denizlg24.com`

    ## Authentication

    All endpoints under `/api/uploads/*`, `/api/folders/*`, and `/api/files/*` require authentication.
    Auth endpoints (`/api/auth/me`, `/api/auth/logout`) also require authentication; login is public.

    Two authentication methods are supported:

    - **Bearer Token (JWT):** Obtained via `POST /api/auth/login`. Pass as `Authorization: Bearer <token>`.
    - **API Key:** Pass as `X-API-Key: <key>` header.

    ## Access Control

    - **Private folders** (`/{userId}/...`): Only the owning user can read/write.
    - **Shared folders** (`/shared/...`): All authenticated users can read. Only the owner or a superuser can modify/delete.

    ## Storage Tiering

    Files are stored on SSD (hot) or HDD (cold) based on disk usage watermarks.
    SSD files preserve directory structure; HDD files use a flat layout keyed by file ID.

    ## TUS Resumable Uploads

    File uploads use the [TUS protocol](https://tus.io/) v1.0.0 with `creation` and `termination` extensions.
    Upload sessions expire after 24 hours. Expired sessions are cleaned up automatically.

    ## Error Format

    All errors follow a consistent shape:
...
    {
...
servers:
...
paths:
...
  /api/files:
    get:
      operationId: listFiles
      summary: List files in a folder
      description: |
        Returns a paginated list of files in the specified folder, ordered by `createdAt` descending.
        The `folderId` query parameter is required.
      tags: [Files]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: folderId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the folder to list files from
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated file list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse_FileSummary"
              example:
                data:
                  - id: 990e8400-e29b-41d4-a716-446655440000
                    filename: report.pdf
                    path: /550e8400-e29b-41d4-a716-446655440000/documents/report.pdf
                    mimeType: application/pdf
                    sizeBytes: 2048000
                    tier: ssd
                    createdAt: "2026-03-01T00:00:00.000Z"
                    updatedAt: "2026-03-01T00:00:00.000Z"
                pagination:
                  page: 1
                  limit: 50
                  total: 1
                  totalPages: 1
        "400":
          description: Missing folderId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: MISSING_FOLDER_ID
                  message: folderId query parameter is required
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/FolderNotFound"
  /api/files/{id}:
    get:
      operationId: getFile
      summary: Get file details
      description: |
        Returns full metadata for a single file, including checksum, storage tier,
        access count, and last access timestamp.
      tags: [Files]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: File details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_FileDetail"
              example:
                data:
                  id: 990e8400-e29b-41d4-a716-446655440000
                  filename: report.pdf
                  path: /550e8400-e29b-41d4-a716-446655440000/documents/report.pdf
                  mimeType: application/pdf
                  sizeBytes: 2048000
                  checksum: a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2
                  tier: ssd
                  lastAccessedAt: "2026-03-10T15:30:00.000Z"
                  accessCount: 42
                  createdAt: "2026-03-01T00:00:00.000Z"
                  updatedAt: "2026-03-01T00:00:00.000Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/FileAccessDenied"
        "404":
          $ref: "#/components/responses/FileNotFound"

    patch:
      operationId: updateFile
      summary: Rename or move a file
      description: |
        Renames and/or moves a file. At least one of `filename` or `folderId` must be provided.

        **Filename normalization:**
        - Converted to snake_case with extension preserved in lowercase
        - Invalid characters are rejected

        For SSD-tier files, the physical file is moved on disk. HDD-tier files use a flat
        layout keyed by file ID, so only the database path is updated.

        If the resulting path is identical to the current path (no actual changes),
        the current file metadata is returned without modification.
      tags: [Files]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFileRequest"
            examples:
              rename:
                summary: Rename file
                value:
                  filename: new_report.pdf
              move:
                summary: Move to different folder
                value:
                  folderId: 880e8400-e29b-41d4-a716-446655440000
              both:
                summary: Rename and move
                value:
                  filename: new_report.pdf
                  folderId: 880e8400-e29b-41d4-a716-446655440000
      responses:
        "200":
          description: File updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_FileUpdated"
              example:
                data:
                  id: 990e8400-e29b-41d4-a716-446655440000
                  filename: new_report.pdf
                  path: /550e8400-e29b-41d4-a716-446655440000/documents/new_report.pdf
                  folderId: 770e8400-e29b-41d4-a716-446655440000
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                nothingToUpdate:
                  summary: No changes provided
                  value:
                    error:
                      code: NOTHING_TO_UPDATE
                      message: Provide filename or folderId
                invalidName:
                  summary: Invalid filename
                  value:
                    error:
                      code: INVALID_NAME
                      message: Name contains invalid characters
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/FileAccessDenied"
        "404":
          description: File or target folder not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                fileNotFound:
                  summary: File not found
                  value:
                    error:
                      code: FILE_NOT_FOUND
                      message: File not found
                folderNotFound:
                  summary: Target folder not found
                  value:
                    error:
                      code: FOLDER_NOT_FOUND
                      message: Target folder not found
        "409":
          description: File already exists at target path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: FILE_EXISTS
                  message: A file already exists at the target path

    delete:
      operationId: deleteFile
      summary: Delete a file
      description: |
        Deletes a file from disk and removes its database record.
        Only the file owner or a superuser can delete a file.
      tags: [Files]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_DeletedId"
              example:
                data:
                  id: 990e8400-e29b-41d4-a716-446655440000
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: ACCESS_DENIED
                  message: You do not have permission to delete this file
        "404":
          $ref: "#/components/responses/FileNotFound"
  /api/files/{id}/download:
    get:
      operationId: downloadFile
      summary: Download a file
      description: |
        Streams the file contents. Supports HTTP range requests for partial content
        (video seeking, download resumption, etc.).

        - Without `?download` query param: `Content-Disposition: inline` (browser preview)
        - With `?download` query param: `Content-Disposition: attachment` (force download)

        Each download increments the file's `accessCount` and updates `lastAccessedAt`
        asynchronously (does not block the response).
      tags: [Files]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
        - name: download
          in: query
          required: false
          schema:
            type: string
          description: |
            If present (any value, including empty), forces `Content-Disposition: attachment`.
            Omit for inline viewing.
        - name: Range
          in: header
          required: false
          schema:
            type: string
            pattern: ^bytes=\d+-\d*$
          description: HTTP range request (e.g., `bytes=0-1023`)
          example: bytes=0-1023
      responses:
        "200":
          description: Full file content
          headers:
            Content-Type:
              schema:
                type: string
              description: File MIME type (falls back to `application/octet-stream`)
            Content-Length:
              schema:
                type: integer
              description: File size in bytes
            Content-Disposition:
              schema:
                type: string
              description: '`inline; filename="..."` or `attachment; filename="..."`'
            Accept-Ranges:
              schema:
                type: string
                const: bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "206":
          description: Partial content (range request)
          headers:
            Content-Type:
              schema:
                type: string
            Content-Length:
              schema:
                type: integer
              description: Size of the returned range
            Content-Range:
              schema:
                type: string
              description: "Range returned (e.g., `bytes 0-1023/4096`)"
            Accept-Ranges:
              schema:
                type: string
                const: bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/FileAccessDenied"
        "404":
          $ref: "#/components/responses/FileNotFound"
        "416":
          description: Range not satisfiable
          headers:
            Content-Range:
              schema:
                type: string
              description: "Total file size (e.g., `bytes */4096`)"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Session JWT obtained via `POST /api/auth/login`.
        Tokens expire after 24 hours by default.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for programmatic access. Created via the admin panel.
        Pass the raw key value (not hashed).
  parameters:
...
    FileId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: File UUID
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (minimum 1)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Items per page (1-100, default 50)
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            noAuth:
              summary: No authentication provided
              value:
                error:
                  code: UNAUTHORIZED
                  message: Authentication required
            tokenInvalid:
              summary: JWT is invalid or expired
              value:
                error:
                  code: TOKEN_INVALID
                  message: Invalid or expired token
            sessionInvalid:
              summary: Session revoked or expired in DB
              value:
                error:
                  code: SESSION_INVALID
                  message: Session expired or revoked
            apiKeyInvalid:
              summary: API key not found
              value:
                error:
                  code: INVALID_API_KEY
                  message: Invalid API key
            apiKeyExpired:
              summary: API key past expiry
              value:
                error:
                  code: API_KEY_EXPIRED
                  message: API key expired
    AccessDenied:
      description: Access denied to the resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: ACCESS_DENIED
              message: You do not have access to this folder
    FolderNotFound:
      description: Folder not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FOLDER_NOT_FOUND
              message: Folder not found
    FileNotFound:
      description: File not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FILE_NOT_FOUND
              message: File not found
    FileAccessDenied:
      description: Access denied to the file
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: ACCESS_DENIED
              message: You do not have access to this file
  schemas:
...
    UpdateFileRequest:
      type: object
      properties:
        filename:
          type: string
          description: New filename (will be normalized to snake_case, extension preserved)
        folderId:
          type: string
          format: uuid
          description: UUID of the target folder (to move the file)

    # --- Response Schemas ---
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
    StorageTier:
      type: string
      enum: [ssd, hdd]
      description: Storage tier where the file resides
...
    FileSummary:
      type: object
      required: [id, filename, path, sizeBytes, tier, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
          description: Normalized filename (snake_case)
        path:
          type: string
          description: Virtual file path
        mimeType:
          type: ["string", "null"]
          description: MIME type (from upload metadata)
        sizeBytes:
          type: integer
          minimum: 0
          description: File size in bytes
        tier:
          $ref: "#/components/schemas/StorageTier"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FileDetail:
      allOf:
        - $ref: "#/components/schemas/FileSummary"
        - type: object
          required: [checksum, accessCount]
          properties:
            checksum:
              type: string
              description: SHA-256 hex digest
              pattern: ^[a-f0-9]{64}$
            lastAccessedAt:
              type: ["string", "null"]
              format: date-time
              description: Timestamp of last download (null if never downloaded)
            accessCount:
              type: integer
              minimum: 0
              description: Number of times the file has been downloaded
...
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
          description: Total number of items
        totalPages:
          type: integer
          minimum: 0
          description: Total number of pages

    # --- Wrapped Response Types ---
...
    ApiResponse_FileDetail:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/FileDetail"
    ApiResponse_FileUpdated:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, filename, path, folderId]
          properties:
            id:
              type: string
              format: uuid
            filename:
              type: string
            path:
              type: string
            folderId:
              type: string
              format: uuid
    ApiResponse_DeletedId:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid
...
    PaginatedResponse_FileSummary:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FileSummary"
        pagination:
          $ref: "#/components/schemas/Pagination"